refactor: move stamp
