+++ packages/components/stamp/src/index.js
