+++ packages/components/stamp/src/stamp.js
