+++ packages/components/stamp/src/stamp.spec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render } from '../../test-utils';
+import { render } from '../../../../src/test-utils';
 import Stamp from './stamp';
 
 it('should render the children', () => {
