+++ packages/components/stamp/src/stamp.story.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import Section from '../../../.storybook/decorators/section';
-import Text from '../typography/text';
-import Spacings from '../spacings';
-import * as icons from '../icons';
+import Section from '../../../../.storybook/decorators/section';
+import Text from '../../../../src/components/typography/text';
+import Spacings from '../../../../src/components/spacings';
+import * as icons from '../../../../src/components/icons';
 import Stamp, { availableTones } from './stamp';
-import Readme from './README.md';
+import Readme from '../README.md';
 
 const iconNames = Object.keys(icons);
 const numberOfIcons = iconNames.length;
