+++ packages/components/stamp/src/stamp.visualspec.js
